hw 1, task 2, subtask I; it should returns long type, avoid type mismatch

+++ A1/SumEven.java
@@ -4,9 +4,9 @@ public class SumEven {
      * The time complexity of this function is O(n), so roughly n steps.
      */
     public static long sumEven(int n) {
-        Integer sum = 0;
+        long sum = 0;
         for (int i = 0; i < n; i++) {
-            sum += i << 1; // bit shift i to the right by 1, same as 2 * i
+            sum += (long)(i << 1); // bit shift i to the right by 1, same as 2 * i
         }
 
         return sum;
